232 add new ringbuffer
